+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - delete() remove await from swarm leave call #209
 - Improve formatting #207
 - Parents validated upon init()
+- tests: clean up #237
 
 ### Changed
 - Location of keys is now `~/.p2pcommons/.hyperdrive` instead of `~/.p2pcommons/.dat`